+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Added:
 - Command-line options for `--log` and `--profile`
 - `--verbose` and `--debug` options set the log level
 - Option to set random seed, with updated tests
+- GP-2093 Bedtools coverage depth metrics
 
 v0.1.6: 2019-08-16
 ------------------